+++ src/index.html
@@ -77,7 +77,7 @@
         </div>
 
     </div>
-    <script src="<%= htmlWebpackPlugin.files.js %>"></script>
+    <script src="<%=htmlWebpackPlugin.files.js%>" defer></script>
     <!--<script src="js/card.js"></script>
     <script src="js/cardList.js"></script>
     <script src="js/formValidator.js"></script>
