+++ webpack.config.js
@@ -28,7 +28,9 @@ module.exports = {
                  'file-loader?name=./images/[name].[ext]',
                  {
                      loader: 'image-webpack-loader',
-                     options: {}
+                     options: {
+                         esModule: false
+                     }
                  },
              ]
          },
